pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meghdoot-ai"
version = "0.1.0"
description = "Latent Diffusion Model for high-resolution weather nowcasting over the Indian subcontinent using INSAT-3DR/3DS satellite imagery"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Vedanschi"},
]

dependencies = [
    # Core Scientific
    "numpy>=1.24,<2.0",
    "scipy>=1.11",
    "pandas>=2.0",

    # Deep Learning
    "torch>=2.1",
    "torchvision>=0.16",
    "diffusers>=0.25",
    "transformers>=4.36",
    "accelerate>=0.25",

    # Geospatial
    "xarray>=2023.10",
    "rasterio>=1.3",
    "h5py>=3.10",
    "netCDF4>=1.6",
    "pyproj>=3.6",
    "opencv-python>=4.8",
    "GDAL>=3.6",

    # Visualization & Monitoring
    "matplotlib>=3.8",
    "plotly>=5.18",
    "wandb>=0.16",

    # Metrics
    "scikit-image>=0.22",
    "scikit-learn>=1.3",

    # Deployment
    "fastapi>=0.108",
    "uvicorn[standard]>=0.25",
    "streamlit>=1.30",
    "folium>=0.15",
    "streamlit-folium>=0.17",
    "docker>=7.0",

    # Utilities
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "mypy>=1.7",
    "pre-commit>=3.6",
    "jupyter>=1.0",
    "ipykernel>=6.27",
]
baselines = [
    "pysteps>=1.8",
    "tensorflow>=2.15",
]

[project.scripts]
meghdoot = "meghdoot.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
